+++ velox/functions/sparksql/CMakeLists.txt
@@ -25,7 +25,6 @@ add_library(
   Register.cpp
   RegisterArithmetic.cpp
   RegisterCompare.cpp
-  RegisterRand.cpp
   Size.cpp
   SplitFunctions.cpp
   String.cpp)
